requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.29.0
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
cryptography==43.0.0
Faker==26.0.0
granian==1.5.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
litestar==2.9.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
multidict==6.0.5
packaging==24.1
polyfactory==2.16.2
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.7
rich==13.7.1
rich-click==1.8.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
types-cffi==1.16.0.20240331
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20240726
types-setuptools==71.1.0.20240726
typing_extensions==4.12.2
uvicorn==0.30.3
uvloop==0.19.0
